+++ source/Classes/TimeSeries.vb
@@ -635,8 +635,8 @@ Public Class TimeSeries
         Dim iStart, iEnd As Integer
         Dim newNodes As SortedList(Of DateTime, Double)
 
-        If _start >= _end Then
-            Throw New ArgumentException("Unable to cut time series: start date is on or after end date!")
+        If _start > _end Then
+            Throw New ArgumentException("Unable to cut time series: start date is after end date!")
         End If
 
         If (Me.StartDate < _start Or Me.EndDate > _end) Then